refactor(srv): split identity_handlers.rs into identity_auth_{spawn,dirs,persist}.rs
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.
Despite the filename, this file was scoped to pre-launch OAuth-flow RPC
handlers only, but bundled four technical layers: RPC dispatch/wire types,
a subprocess spawn/drain engine with near-duplicate pipes-vs-PTY variants
(~670 lines, nearly half the file), filesystem/directory provisioning, and
DB persistence. Split into identity_handlers.rs (dispatch/wire types,
register_identity_handlers) + identity_auth_spawn.rs + identity_auth_dirs.rs
+ identity_auth_persist.rs, tests colocated with the code each exercises.
compute_and_ensure_bundle_dir (flagged vestigial in the audit) was moved,
not deleted — deletion is a separate decision.

Pure relocation, no logic changes. cargo build -p agentmux-srv clean;
16/16 tests pass across the 4 files, matching the original combined count.

<!-- agentmux:agent_id=agentx -->
