+++ agentmux-srv/src/server/mod.rs
@@ -6,6 +6,9 @@ mod files;
 pub(crate) mod app_api;
 mod agent_handlers;
 mod editor_handlers;
+mod identity_auth_dirs;
+mod identity_auth_persist;
+mod identity_auth_spawn;
 mod identity_handlers;
 pub mod install_handlers;
 mod lsp_handlers;
